requirements.txt
anthropic==0.75.0
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
gunicorn==23.0.0
langdetect==1.0.9
numpy>=1.26.0,<2.0.0
opencv-python-headless==4.10.0.84
pandas==2.2.3
psycopg2-binary==2.9.11
PyMuPDF==1.24.14
PyPDF2==3.0.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.2.0
pytz==2025.2
requests==2.32.5
Werkzeug==3.1.3
Jinja2==3.1.6
MarkupSafe==3.0.3
itsdangerous==2.2.0
click==8.3.1
regex==2025.11.3
pdf2image==1.17.0
pillow==12.0.0
urllib3==2.5.0
python-dotenv==1.2.1

